Create command clear and create file configuration messages

+++ src/app/commands/clear.js
@@ -1,3 +1,5 @@
+import configuration from '../../../configure';
+
 class Clear {
 	constructor() {
 		this.config = {
@@ -9,19 +11,33 @@ class Clear {
 		};
 		this.run = async (_, msg, args, prefix) => {
 			const n = Number(args[0]);
-			if (!n || n < 1 || n > 100)
-				return msg.reply(
+			if (!n || n < 1 || n > 100 || args[0]) {
+				msg.reply(
 					`Sintaxe incorreta por favor digite assim: \`${
 						prefix + this.config.name
-					} {1/100}\``
+					} {1/100/all}\``
 				);
+				return `O usuário mencionou um número inválido para a limpeza.`;
+			}
 			try {
 				await msg.channel.bulkDelete(n);
-				msg.channel.send(`Você limpou \`${n}\` mensagens.`);
-				return true;
+				msg.channel.send(
+					configuration.comandos.clear.apagouMensagens
+						.replace('$MESSAGES_DELETED', n)
+						.replace('$USERNAME', msg.member.user.username)
+						.replace('$USER_TAG', msg.member.user.discriminator)
+				);
+				return `Com esse comando o usuário apagou ${n} mensagens, no canal ${msg.channel.id}.`;
 			} catch (err) {
-				msg.reply('Houve um erro para deletar as mensagens.');
-				return false;
+				console.log(err);
+				msg.reply(
+					configuration.comandos.clear.apagouMensagens
+						.replace('$MESSAGES_DELETED', n)
+						.replace('$USERNAME', msg.member.user.username)
+						.replace('$USER_TAG', msg.member.user.discriminator)
+						.replace('$ERROR_MESSAGE', err.message)
+				);
+				return `Houve um erro ao usuário executar esse comando.`;
 			}
 		};
 	}
